generic checks for payment, paymentchannelclaim, escrowfinish credential list
