+++ packages/xrpl/src/models/transactions/escrowFinish.ts
@@ -5,6 +5,7 @@ import {
   BaseTransaction,
   isAccount,
   validateBaseTransaction,
+  validateCredentialsList,
   validateRequiredField,
 } from './common'
 
@@ -48,6 +49,11 @@ export function validateEscrowFinish(tx: Record<string, unknown>): void {
   validateBaseTransaction(tx)
 
   validateRequiredField(tx, 'Owner', isAccount)
+
+  if (tx.CredentialIDs !== undefined) {
+    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- known from base check
+    validateCredentialsList(tx.CredentialIDs, tx.TransactionType as string)
+  }
 
   if (tx.OfferSequence == null) {
     throw new ValidationError('EscrowFinish: missing field OfferSequence')
